requirements.txt
fastapi
uvicorn
pandas
langchain
openai
python-dotenv
ollama
langchain-community
pinecone
pyreadline3
langchain-ollama
sentence-transformers
tqdm